+++ src/Application/RoadGeocoderInterface.php
@@ -7,4 +7,6 @@ namespace App\Application;
 interface RoadGeocoderInterface
 {
     public function computeRoadLine(string $roadName, string $inseeCode): string;
+
+    public function getDepartmentalRoad(string $search, string $gestionnaire, string $roadType): array;
 }
