+++ src/Infrastructure/Controller/Regulation/Fragments/GetDepartmentalRoadCompletionFragmentController.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Infrastructure\Controller\Regulation\Fragments;
 
-use App\Application\RoadsNumbersInterface;
+use App\Application\RoadGeocoderInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
@@ -14,7 +14,7 @@ final class GetDepartmentalRoadCompletionFragmentController
 {
     public function __construct(
         private \Twig\Environment $twig,
-        private RoadsNumbersInterface $roadsNumbers,
+        private RoadGeocoderInterface $roadGeocoder,
     ) {
     }
 
@@ -33,7 +33,7 @@ final class GetDepartmentalRoadCompletionFragmentController
         }
 
         $roadType = 'Départementale';
-        $departmentalRoadsNumbers = $this->roadsNumbers->getDepartmentalRoad($search, $administrator, $roadType);
+        $departmentalRoadsNumbers = $this->roadGeocoder->getDepartmentalRoad($search, $administrator, $roadType);
 
         return new Response(
             $this->twig->render(
