+++ tests/Unit/Infrastructure/Adapter/IgnWfsGeocoderTest.php
@@ -5,18 +5,39 @@ declare(strict_types=1);
 namespace App\Tests\Unit\Infrastructure\Adapter;
 
 use App\Application\Exception\GeocodingFailureException;
-use App\Infrastructure\Adapter\IgnWfsRoadsNumbers;
+use App\Infrastructure\Adapter\IgnWfsGeocoder;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\HttpClient\MockHttpClient;
 use Symfony\Component\HttpClient\Response\MockResponse;
 
-final class IgnWfsRoadsNumbersTest extends TestCase
+final class IgnWfsGeocoderTest extends TestCase
 {
+    private string $address = 'Rue Saint-Victor 59110 La Madeleine';
+    private string $inseeCode = '59368';
     private string $departmentalRoadNumber = 'D32';
     private string $gestionnaire = 'Ardennes';
     private string $roadType = 'Départementale';
     private string $baseUrl = 'http://testserver';
     private string $ignWfsUrl = 'http://testserver/wfs/ows';
+
+    public function testComputeRoadLine(): void
+    {
+        $body = '{"features": [{"geometry": {"type": "MultiLineString", "coordinates": ["..."]}}]}';
+        $response = new MockResponse($body, ['http_code' => 200]);
+        $http = new MockHttpClient([$response], $this->baseUrl);
+
+        $roadGeocoder = new IgnWfsGeocoder($this->ignWfsUrl, $http);
+
+        $geometry = $roadGeocoder->computeRoadLine($this->address, $this->inseeCode);
+
+        $this->assertSame('{"type":"MultiLineString","coordinates":["..."]}', $geometry);
+
+        $this->assertSame('GET', $response->getRequestMethod());
+        $this->assertSame(
+            'http://testserver/wfs/ows?SERVICE=WFS&REQUEST=GetFeature&VERSION=2.0.0&OUTPUTFORMAT=application/json&TYPENAME=BDTOPO_V3:voie_nommee&cql_filter=strStripAccents(nom_minuscule)%3DstrStripAccents(%27rue%20saint-victor%2059110%20la%20madeleine%27)%20AND%20code_insee%3D%2759368%27&PropertyName=geometrie',
+            $response->getRequestUrl(),
+        );
+    }
 
     public function testGetDepartmentalRoad(): void
     {
@@ -25,9 +46,9 @@ final class IgnWfsRoadsNumbersTest extends TestCase
         $response = new MockResponse($body, ['http_code' => 200]);
         $http = new MockHttpClient([$response], $this->baseUrl);
 
-        $IgnWfsRoadsNumbers = new IgnWfsRoadsNumbers($this->ignWfsUrl, $http);
+        $IgnWfsGeocoder = new IgnWfsGeocoder($this->ignWfsUrl, $http);
 
-        $geometry = $IgnWfsRoadsNumbers->getDepartmentalRoad($this->departmentalRoadNumber, $this->gestionnaire, $this->roadType);
+        $geometry = $IgnWfsGeocoder->getDepartmentalRoad($this->departmentalRoadNumber, $this->gestionnaire, $this->roadType);
         $this->assertSame('D32', $geometry[0]['numero']);
         $this->assertSame('{"type":"MultiLineString","coordinates":[]}', $geometry[0]['geometry']);
         $this->assertSame('GET', $response->getRequestMethod());
@@ -57,7 +78,7 @@ final class IgnWfsRoadsNumbersTest extends TestCase
         $response = new MockResponse('...', ['http_code' => $statusCode]);
         $http = new MockHttpClient([$response]);
 
-        $roadGeocoder = new IgnWfsRoadGeocoder($this->ignWfsUrl, $http);
+        $roadGeocoder = new IgnWfsGeocoder($this->ignWfsUrl, $http);
         $roadGeocoder->computeRoadLine($this->address, $this->inseeCode);
     }
 
@@ -82,7 +103,31 @@ final class IgnWfsRoadsNumbersTest extends TestCase
         $response = new MockResponse($body, ['http_code' => 200]);
         $http = new MockHttpClient([$response]);
 
-        $IgnWfsRoadsNumbers = new IgnWfsRoadsNumbers($this->ignWfsUrl, $http);
-        $IgnWfsRoadsNumbers->getDepartmentalRoad($this->departmentalRoadNumber, $this->gestionnaire, $this->roadType);
+        $IgnWfsGeocoder = new IgnWfsGeocoder($this->ignWfsUrl, $http);
+
+        $IgnWfsGeocoder->computeRoadLine($this->address, $this->inseeCode);
+    }
+
+    private function provideDecodeErrorDataForDepartmental(): array
+    {
+        return [
+            ['{"features": }', '/invalid json: Syntax error/'],
+        ];
+    }
+
+    /**
+     * @dataProvider provideDecodeErrorDataForDepartmental
+     */
+    public function testGetDepartmentalRoadError(string $body, string $pattern): void
+    {
+        $this->expectException(GeocodingFailureException::class);
+        $this->expectExceptionMessageMatches($pattern);
+
+        $response = new MockResponse($body, ['http_code' => 200]);
+        $http = new MockHttpClient([$response]);
+
+        $IgnWfsGeocoder = new IgnWfsGeocoder($this->ignWfsUrl, $http);
+
+        $IgnWfsGeocoder->getDepartmentalRoad($this->departmentalRoadNumber, $this->gestionnaire, $this->roadType);
     }
 }
